+++ docs/src/markdown/about/changelog.md
@@ -4,6 +4,7 @@
 
 - **NEW**: Formally add support for Python 3.11.
 - **NEW**: `set()` can now take a dictionary of channels and values and set multiple channels at once.
+- **NEW**: `get()` can now take a list of channels and will return a list of those channel values.
 - **ENHANCE**: Simplify some type annotation syntax.
 - **ENHANCE**: Some minor performance enhancements.
 - **FIX**: Fix OkLCh CSS parsing.